backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
scikit-learn==1.6.1
numpy==2.2.4
pandas==2.2.3
scipy==1.15.2
joblib==1.4.2
threadpoolctl==3.6.0
python-dotenv==1.0.1 